refactor(sdk): merge ERC-1363 tests into shield.test.ts [SDK-145]
